release: v0.9.0 (#981)

+++ test/e2e/testenv_test.go
@@ -12,11 +12,12 @@ import (
 
 	"github.com/hashicorp/go-hclog"
 	"github.com/pkg/errors"
-	"github.com/sarvalabs/battleground/network/infrastructure"
-	"github.com/sarvalabs/battleground/types"
-	"github.com/sarvalabs/go-moi-identifiers"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
+
+	"github.com/sarvalabs/battleground/network/infrastructure"
+	"github.com/sarvalabs/battleground/types"
+	identifiers "github.com/sarvalabs/go-moi-identifiers"
 
 	"github.com/sarvalabs/go-moi/bgclient"
 	cmdcommon "github.com/sarvalabs/go-moi/cmd/common"
@@ -41,7 +42,7 @@ const (
 	DefaultShutdownTimeout  = 10 * time.Minute
 	DefaultConfirmIxTimeout = 1 * time.Minute
 	DefaultAccountCount     = 2 // 2 accounts are enough to fire ixns in debug mode
-	InitialKMOITokens       = 25000
+	InitialKMOITokens       = 50000
 	DefaultJSONRPCPort      = 29000
 )
 
@@ -57,6 +58,10 @@ var (
 	commitHash       = flag.String("commit-hash", "", "enter the commit hash of the repo to be deployed")
 	cleanDB          = flag.Bool("clean-db", true, "specify whether to clean the database")
 	logLevel         = flag.String("log-level", "TRACE", "enter the log level")
+	debug            = flag.Bool("debug", false, "specify whether to collect logs")
+	shouldExecute    = flag.Bool("should-execute", true, "specify whether to execute tesseract while syncing")
+	oldState         = flag.Bool("old-state", false, "specify whether to run on previous state")
+	enableSorition   = flag.Bool("enable-sortition", false, "specify whether to use sortition for operator selection")
 )
 
 type BattleGroundConfig struct {
@@ -139,13 +144,16 @@ func (te *TestEnvironment) configureBattleGround() error {
 		te.validatorCount = bgConfig.NoOfInstances * bgConfig.NoOfPodsPerInstance
 	} else {
 		d := bgclient.DefaultClusterConfig()
-		d.WithLogs = false
+		d.WithLogs = *debug
 		d.WithStdout = false
 		d.LogLevel = "TRACE"
 		d.BootNodePort = 27000
 		d.Libp2pPort = 28000
 		d.JSONRPCPort = DefaultJSONRPCPort
 		d.GuardianPath = "../../moiclient/"
+		d.ShouldExecute = *shouldExecute
+		d.OldState = *oldState
+		d.EnableSortition = *enableSorition
 
 		te.bgClient = bgclient.NewClient(&bgclient.Config{
 			ClusterConfig: d,
@@ -233,18 +241,26 @@ func (te *TestEnvironment) initializeMOIClient() {
 }
 
 func (te *TestEnvironment) runCriticallyNecessaryTearDown() {
-	ctx, cancel := context.WithTimeout(context.Background(), DefaultShutdownTimeout)
-	defer cancel()
+	if !(*debug) {
+		ctx, cancel := context.WithTimeout(context.Background(), DefaultShutdownTimeout)
+		defer cancel()
 
-	err := te.bgClient.DestroyNetwork(ctx, true)
-	te.Suite.NoError(err)
+		err := te.bgClient.DestroyNetwork(ctx, true)
+		te.Suite.NoError(err)
+	}
 }
 
 func (te *TestEnvironment) initLogger() {
-	te.logger = hclog.New(&hclog.LoggerOptions{
-		Name:  "E2E",
-		Level: hclog.LevelFromString(*logLevel),
-	})
+	if *debug {
+		te.logger = hclog.New(&hclog.LoggerOptions{
+			Name:  "E2E",
+			Level: hclog.LevelFromString(*logLevel),
+		})
+
+		return
+	}
+
+	te.logger = hclog.NewNullLogger()
 }
 
 // fetch json urls and check if initial sync is done in all nodes
